% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineEstimates.R
\name{combineEstimates}
\alias{combineEstimates}
\title{Combine estimates}
\usage{
combineEstimates(parentFolder, cmFolders, maxCores = 1)
}
\arguments{
\item{cmFolders}{Name of local folders where the CohortMethod runs can be found.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}
}
\description{
Combine estimates
}
