% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePerformance.R
\name{computePerformance}
\alias{computePerformance}
\title{Compute performance metrics of a run}
\usage{
computePerformance(
  outputFolder,
  cmFolder,
  maxCores = 1,
  outputFileName,
  databaseId
)
}
\arguments{
\item{outputFolder}{Name of local folder to place intermediary results; make sure to use
forward slashes (/). Do not use a folder on a network drive since}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{referenceSet}{The name of the reference set for which outcomes need to be created.
Currently supported are "ohdsiMethodsBenchmark", and "ohdsiDevelopment".}

\item{cmFolders}{Name of local folders where the CohortMethod runs can be found.}
}
\description{
Compute performance metrics of a run
}
