% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallSampleEstimationEvaluation.R
\docType{package}
\name{SmallSampleEstimationEvaluation-package}
\alias{SmallSampleEstimationEvaluation}
\alias{SmallSampleEstimationEvaluation-package}
\title{SmallSampleEstimationEvaluation: Package for Evaluation of Estimation Methods}
\description{
Evaluation of the performance of comparative effect estimation methods when sample size (number of subjects in the target and comparator) is small.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ohdsi-studies/SmallSampleEstimationEvaluation/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
